Move interface declaration and map instantiation to provider.go
